log data to console, films returning, but cinemas are not
