+++ Lib/getAllCinemas.ts
@@ -2,13 +2,14 @@ import { CinemaType } from "@/Types/Object-Interfaces";
 import supabaseClient from "./supabaseClient";
 
 const getAllCinemas = async (): Promise<CinemaType[]> => {
-  const { data, error } = await supabaseClient.from("venue").select("*");
+  let { data, error } = await supabaseClient.from("venue").select("*");
 
   if (error) {
     console.log(`Error getting cinemas: ${error.message}`);
     throw error;
   }
 
+  console.log("Data :", data);
   return data;
 };
